% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{as_neurovol}
\alias{as_neurovol}
\title{Memoised version of .as_neurovol_unmemoised
This is the function that should be called by get_template}
\usage{
as_neurovol(path_or_py_obj)
}
\description{
Memoised version of .as_neurovol_unmemoised
This is the function that should be called by get_template
}
\keyword{internal}
