% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_ref.R
\name{atlas_ref}
\alias{atlas_ref}
\alias{atlas_ref.atlas}
\alias{atlas_ref.default}
\title{Atlas Reference Accessor}
\usage{
atlas_ref(x, ...)

\method{atlas_ref}{atlas}(x, ...)

\method{atlas_ref}{default}(x, ...)
}
\arguments{
\item{x}{An atlas object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class `"atlas_ref"`.
}
\description{
Returns the canonical atlas reference metadata for an atlas object.
}
\examples{
grid <- neuroim2::NeuroSpace(c(2, 1, 1))
volume <- neuroim2::NeuroVol(array(c(1, 2), c(2, 1, 1)), grid)
atlas <- structure(
  list(
    atlas = volume, ids = 1:2,
    labels = c("A", "B"), orig_labels = c("A", "B"),
    hemi = c("left", "right"), name = "toy",
    cmap = rbind(c(1, 0, 0), c(0, 0, 1)),
    atlas_ref = new_atlas_ref("toy", "two-regions",
      template_space = "MNI152NLin6Asym", coord_space = "MNI152"
    )
  ),
  class = "atlas"
)
atlas_ref(atlas)
}
