% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{$.templateflow}
\alias{$.templateflow}
\title{Access Attributes of the TemplateFlow Object (DEPRECATED)}
\usage{
\method{$}{templateflow}(x, name)
}
\arguments{
\item{x}{An object of class \code{templateflow}.}

\item{name}{The name of the attribute to access.}
}
\value{
The list element, or NULL if not found.
}
\description{
Access Attributes of the TemplateFlow Object (DEPRECATED)
}
\keyword{internal}
