% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_templateflow}
\alias{check_templateflow}
\title{Check TemplateFlow Installation Status}
\usage{
check_templateflow()
}
\value{
Invisible NULL. Prints status information to the console.
}
\description{
Checks whether the TemplateFlow R package is installed and provides
cache information.
}
\examples{
\donttest{
check_templateflow()
}
}
