% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\name{.schaefer_templateflow_surf}
\alias{.schaefer_templateflow_surf}
\title{Schaefer surface atlas (TemplateFlow-backed helper)}
\usage{
.schaefer_templateflow_surf(parcels, networks, space, surf, use_cache = TRUE)
}
\description{
This helper uses TemplateFlow for surface geometry and CBIG for Schaefer
annotation (.annot) files. It is used for spaces other than fsaverage6.
}
\keyword{internal}
