% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsaverage.R
\docType{data}
\name{fsaverage}
\alias{fsaverage}
\title{Surface geometry for the fsaverage6 atlas}
\format{
A named list with elements \code{lh} and \code{rh}, each containing
  surface geometry objects for orig, white, inflated, and pial surfaces.
}
\usage{
data(fsaverage)
}
\description{
A list including left and right hemispheres for the orig, white, inflated,
and pial surfaces.
}
\examples{
\donttest{
data(fsaverage)
names(fsaverage)
}

}
\keyword{datasets}
