% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/atlas_methods.R,
%   R/ggschaefer.R, R/glasser.R
\name{map_atlas}
\alias{map_atlas}
\alias{map_atlas.atlas}
\alias{map_atlas.schaefer}
\alias{map_atlas.glasser}
\title{Map values to an atlas}
\usage{
map_atlas(x, vals, thresh, ...)

\method{map_atlas}{atlas}(x, vals, thresh = NULL, pos = FALSE, ...)

\method{map_atlas}{schaefer}(x, vals, thresh = NULL, pos = FALSE, ...)

\method{map_atlas}{glasser}(x, vals, thresh = NULL, pos = FALSE, ...)
}
\arguments{
\item{x}{An atlas object to map values onto}

\item{vals}{Numeric vector of values to map to atlas regions. Length should
match
the number of regions in the atlas}

\item{thresh}{Optional numeric vector of length 2 specifying (min, max)
thresholds
for the mapped values. Values outside this range will be clamped.}

\item{...}{Additional arguments passed to methods}

\item{pos}{Logical. If `TRUE`, values are thresholded using raw values;
otherwise the absolute values are used.}
}
\value{
Returns the atlas object with mapped values
}
\description{
Maps a set of values to regions/parcels in an atlas object. This can be used
to visualize data (like statistics or measurements) across atlas regions.
}
\examples{
\dontrun{
# Load the aseg atlas
atlas <- get_aseg_atlas()
vals <- rnorm(length(atlas$orig_labels))

# Map values with a threshold of -2 to 2
mapped <- map_atlas(atlas, vals, thresh = c(-2, 2))
}
}
