% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{names.templateflow}
\alias{names.templateflow}
\title{List Attributes of the TemplateFlow Object (DEPRECATED)}
\usage{
\method{names}{templateflow}(x)
}
\arguments{
\item{x}{An object of class \code{templateflow}.}
}
\value{
A character vector of available element names.
}
\description{
List Attributes of the TemplateFlow Object (DEPRECATED)
}
\keyword{internal}
