% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R, R/atlas_provenance.R, R/glasser.R,
%   R/print-plot-methods.R
\name{print.atlas}
\alias{print.atlas}
\alias{print.atlas_provenance}
\alias{print.glasser}
\alias{print-methods}
\alias{print.schaefer}
\title{Print Methods for neuroatlas Objects}
\usage{
\method{print}{atlas}(x, ...)

\method{print}{atlas_provenance}(x, ...)

\method{print}{glasser}(x, ...)

\method{print}{schaefer}(x, ...)
}
\arguments{
\item{x}{An atlas object (atlas, glasser, schaefer, etc.)}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns \code{x}.

The object is returned invisibly
}
\description{
Print methods for various atlas objects in the neuroatlas package
}
\examples{
atlas <- get_aseg_atlas()
print(atlas)

}
