% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_ref.R
\name{print.atlas_ref}
\alias{print.atlas_ref}
\title{Print Method for Atlas References}
\usage{
\method{print}{atlas_ref}(x, ...)
}
\arguments{
\item{x}{An `atlas_ref` object.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print Method for Atlas References
}
\examples{
ref <- new_atlas_ref("toy", "two-regions",
  template_space = "MNI152NLin6Asym", coord_space = "MNI152"
)
print(ref)
}
