% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_registry.R
\name{print.atlas_spec}
\alias{print.atlas_spec}
\title{Print an atlas_spec}
\usage{
\method{print}{atlas_spec}(x, ...)
}
\arguments{
\item{x}{An `atlas_spec` object.}

\item{...}{Unused.}
}
\description{
Print an atlas_spec
}
\keyword{internal}
