% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_ref.R
\name{validate_atlas_ref}
\alias{validate_atlas_ref}
\title{Validate an Atlas Reference}
\usage{
validate_atlas_ref(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
Invisibly returns `x` when valid.
}
\description{
Validate an Atlas Reference
}
\examples{
ref <- new_atlas_ref("toy", "two-regions",
  template_space = "MNI152NLin6Asym", coord_space = "MNI152"
)
validate_atlas_ref(ref)
}
