% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_ref.R
\name{atlas_coord_space}
\alias{atlas_coord_space}
\title{Atlas Coordinate-Space Convenience Accessor}
\usage{
atlas_coord_space(x)
}
\arguments{
\item{x}{An atlas object.}
}
\value{
Character scalar (or `NA_character_`).
}
\description{
Atlas Coordinate-Space Convenience Accessor
}
