% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_ref.R
\name{atlas_family}
\alias{atlas_family}
\title{Atlas Family Convenience Accessor}
\usage{
atlas_family(x)
}
\arguments{
\item{x}{An atlas object.}
}
\value{
Character scalar.
}
\description{
Atlas Family Convenience Accessor
}
