% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_ref.R
\name{atlas_ref}
\alias{atlas_ref}
\alias{atlas_ref.atlas}
\alias{atlas_ref.default}
\title{Atlas Reference Accessor}
\usage{
atlas_ref(x, ...)

\method{atlas_ref}{atlas}(x, ...)

\method{atlas_ref}{default}(x, ...)
}
\arguments{
\item{x}{An atlas object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class `"atlas_ref"`.
}
\description{
Returns the canonical atlas reference metadata for an atlas object.
}
