% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_ref.R
\name{atlas_space}
\alias{atlas_space}
\title{Atlas Template-Space Convenience Accessor}
\usage{
atlas_space(x)
}
\arguments{
\item{x}{An atlas object.}
}
\value{
Character scalar (or `NA_character_`).
}
\description{
Atlas Template-Space Convenience Accessor
}
