% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_spaces.R
\docType{data}
\name{coord_spaces}
\alias{coord_spaces}
\title{Standard Coordinate Space Identifiers}
\format{
A named list with the following elements:
\describe{
  \item{MNI305}{FreeSurfer/Talairach space (fsaverage native)}
  \item{MNI152}{ICBM 2009c space (common fMRI template)}
  \item{SCANNER}{Native scanner space (subject-specific)}
  \item{UNKNOWN}{Unknown or unspecified space}
}
}
\usage{
coord_spaces
}
\description{
Character constants for standard neuroimaging coordinate spaces.
}
\examples{
coord_spaces$MNI305
coord_spaces$MNI152

}
\keyword{datasets}
