% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilate_parcels.R
\name{dilate_atlas}
\alias{dilate_atlas}
\title{Dilate the Boundaries of an Atlas Parcellation within a Mask}
\usage{
dilate_atlas(atlas, mask, radius = 4, maxn = 50)
}
\arguments{
\item{atlas}{A brain atlas object. Must have the class "atlas".}

\item{mask}{A mask of voxels to include for potential dilation.}

\item{radius}{The dilation expansion radius (default: 4).}

\item{maxn}{The maximum number of neighbors to consider for dilation (default: 50).}
}
\value{
A ClusteredNeuroVol object representing the dilated atlas parcellation.
}
\description{
This function dilates the boundaries of an atlas parcellation within a specified mask,
expanding the parcellation by a given radius.
}
\examples{
atlas <- get_aseg_atlas()
mask <- create_brain_mask(atlas)
dilated_atlas <- dilate_atlas(atlas, mask, radius = 4, maxn = 50)
}
