% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilate_parcels.R
\name{dilate_atlas}
\alias{dilate_atlas}
\title{dilate the boundaries of an atlas parcellation within a mask}
\usage{
dilate_atlas(atlas, mask, radius = 4, maxn = 50)
}
\arguments{
\item{atlas}{the brain atlas}

\item{mask}{the mask of voxels to include for potential dilation}

\item{radius}{the dilation expansion radius}

\item{maxn}{the maximum number of neighbors to consider for dilation}
}
\description{
dilate the boundaries of an atlas parcellation within a mask
}
