% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{.as_neurovol_unmemoised}
\alias{.as_neurovol_unmemoised}
\title{Internal Helper to Convert File Path to NeuroVol}
\usage{
.as_neurovol_unmemoised(file_path)
}
\arguments{
\item{file_path}{A character string file path to a NIfTI file.}
}
\value{
A \code{neuroim2::NeuroVol} object.
}
\description{
Reads a NIfTI file into a \code{neuroim2::NeuroVol} object.
}
\keyword{internal}
