% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{.neuroatlas_cache_dir}
\alias{.neuroatlas_cache_dir}
\title{Get or Create neuroatlas Cache Directory}
\usage{
.neuroatlas_cache_dir(subdir = NULL)
}
\arguments{
\item{subdir}{Optional character string. If provided, a subdirectory named
`subdir` will be created/used within the main neuroatlas cache directory.}
}
\value{
A character string representing the path to the cache directory.
}
\description{
Returns a path to a neuroatlas-specific cache directory. If the directory
(or a specified subdirectory) doesn't exist, it will be created.
This function uses `tools::R_user_dir` to ensure a user-specific,
OS-appropriate cache location.
}
\keyword{internal}
