% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsaverage.R
\docType{data}
\name{fsaverage}
\alias{fsaverage}
\title{Surface geometry for the fsaverage6 atlas}
\format{
An object of class \code{list} of length 7.
}
\usage{
data(fsaverage)
}
\description{
a list including left and right hemispheres for the orig, white, inflated, and pial surfaces.
}
\keyword{datasets}
