% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasser.R
\name{get_glasser_atlas}
\alias{get_glasser_atlas}
\title{Retrieve and load Glasser360 network parcellation from PennBBL github repository}
\usage{
get_glasser_atlas(outspace = NULL)
}
\arguments{
\item{parcels}{the number rof parcels in hte atlas}

\item{networks}{the network number}

\item{resolution}{the resolution of the atlas in MNI space (1mm or 2mm)}

\item{outdim}{an optional 3-element vector indicating dimensions to resample atlas to.}
}
\value{
a list with three elemenents:

\describe{
  \item{atlas}{ a NeuroVol instance with integer indices coding the ROIs}
  \item{cmap}{ a data.frame indicating the colors associated with each ROI}
  \item{legend}{a data.frame providing label information about each ROI}
}
}
\description{
Retrieve and load Glasser360 network parcellation from PennBBL github repository
}
\details{
Files are downloaded from the github repository: https://github.com/PennBBL/xcpEngine/
}
\examples{

v1 <- get_glasser_atlas()


}
