% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{get_template_brainmask}
\alias{get_template_brainmask}
\title{Get Brain Mask from Template}
\usage{
get_template_brainmask(
  name = "MNI152NLin2009cAsym",
  resolution = 1,
  extension = ".nii.gz"
)
}
\arguments{
\item{name}{Character string specifying template name. Default: "MNI152NLin2009cAsym"}

\item{resolution}{Numeric resolution in mm. Default: 1}

\item{extension}{Character string specifying file extension. Default: ".nii.gz"}
}
\value{
A NeuroVol object containing the binary brain mask
}
\description{
Convenience function to retrieve a binary brain mask for a specified template.
}
