% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{get_template_gm}
\alias{get_template_gm}
\title{Get Gray Matter Probability Map}
\usage{
get_template_gm(
  name = "MNI152NLin2009cAsym",
  resolution = 1,
  extension = ".nii.gz"
)
}
\arguments{
\item{name}{Character string specifying template name. Default: "MNI152NLin2009cAsym"}

\item{resolution}{Numeric resolution in mm. Default: 1}

\item{extension}{Character string specifying file extension. Default: ".nii.gz"}
}
\value{
A NeuroVol object containing the gray matter probability map
}
\description{
Convenience function to get gray matter probability map.
}
