% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_explorer.R
\name{infer_design_var_type}
\alias{infer_design_var_type}
\title{Infer Design Variable Type}
\usage{
infer_design_var_type(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A character scalar.
}
\description{
Classify a vector as \code{"continuous"} or \code{"categorical"} for plot
defaults.
}
