% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_templateflow}
\alias{install_templateflow}
\title{Install Templateflow Python package}
\usage{
install_templateflow(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method passed to \code{reticulate::py_install}.}
\item{conda}{Path to a conda executable or \code{"auto"}.}
}
\value{
Invisible \code{NULL} indicating success
}
\description{
Installs the Python packages \pkg{scipy} and \pkg{templateflow} using \code{reticulate::py_install}.
}
\examples{
\donttest{
install_templateflow()
}
}

