% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_explorer.R
\name{launch_cluster_explorer}
\alias{launch_cluster_explorer}
\title{Launch Cluster Explorer in Interactive Session}
\usage{
launch_cluster_explorer(...)
}
\arguments{
\item{...}{Passed to \code{\link{cluster_explorer}()}.}
}
\value{
Invisibly returns the running app object.
}
\description{
Convenience wrapper around \code{\link{cluster_explorer}()} that launches the
app in an interactive session.
}
