% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{map_atlas}
\alias{map_atlas}
\title{map a set of values to an atlas}
\usage{
map_atlas(x, vals, thresh, ...)
}
\arguments{
\item{x}{the atlas to map values to}

\item{vals}{the values to map}

\item{thresh}{optional (min,max) threshold vector}
}
\description{
map a set of values to an atlas
}
