% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{map_atlas}
\alias{map_atlas}
\alias{map_atlas.atlas}
\alias{map_atlas.glasser}
\title{Map values to an atlas}
\usage{
map_atlas(x, vals, thresh, ...)
}
\arguments{
\item{x}{An atlas object to map values onto}

\item{vals}{Numeric vector of values to map to atlas regions. Length should match
the number of regions in the atlas}

\item{thresh}{Optional numeric vector of length 2 specifying (min, max) thresholds
for the mapped values. Values outside this range will be clamped.}

\item{...}{Additional arguments passed to methods}
}
\value{
Returns the atlas object with mapped values
}
\description{
Maps a set of values to regions/parcels in an atlas object. This can be used
to visualize data (like statistics or measurements) across atlas regions.
}
