% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{merge_atlases}
\alias{merge_atlases}
\title{Merge Atlases}
\usage{
merge_atlases(atlas1, atlas2)
}
\arguments{
\item{atlas1}{The first atlas object to merge.}

\item{atlas2}{The second atlas object to merge.}
}
\value{
A list containing the merged atlas information:
  \itemize{
    \item{name}{The concatenated names of the input atlases, separated by "::".}
    \item{atlas}{A ClusteredNeuroVol object representing the merged atlas.}
    \item{cmap}{A matrix containing the color maps of both atlases.}
    \item{ids}{A vector containing the combined ids of both atlases.}
    \item{labels}{A vector containing the combined labels of both atlases.}
    \item{orig_labels}{A vector containing the combined original labels of both atlases.}
    \item{hemi}{A vector containing the combined hemispheres of both atlases.}
  }
}
\description{
This function merges two atlases into a single atlas object, combining their
labels, color maps, ids, and hemispheres.
}
\examples{
atlas1 <- get_aseg_atlas()
atlas2 <- get_aseg_atlas()
merged_atlas <- merge_atlases(atlas1, atlas2)
}
