% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_spaces.R
\docType{data}
\name{MNI305_to_MNI152}
\alias{MNI305_to_MNI152}
\title{MNI305 to MNI152 Affine Transform Matrix}
\format{
A 4x4 numeric matrix (affine transform in homogeneous coordinates).
}
\source{
FreeSurfer: \code{$FREESURFER_HOME/average/mni152.register.dat}
}
\usage{
MNI305_to_MNI152
}
\description{
The canonical 4x4 affine transformation matrix for converting RAS coordinates
from MNI305 (fsaverage) space to MNI152 space.
}
\details{
This matrix is derived from FreeSurfer's \code{mni152.register.dat} file,
located at \code{$FREESURFER_HOME/average/mni152.register.dat}.

The transform accounts for the approximately 4mm difference between MNI305
and MNI152 coordinate systems. It includes small rotation, scaling, and
translation components.

To apply: for a point \code{p = c(R, A, S)}, compute
\code{MNI305_to_MNI152 \%*\% c(p, 1)} and take the first 3 elements.
}
\examples{
# Transform a single point from MNI305 to MNI152
point_305 <- c(10, -20, 35)
point_152 <- (MNI305_to_MNI152 \%*\% c(point_305, 1))[1:3]
print(point_152)  # approximately c(10.695, -18.409, 36.137)

}
\seealso{
\url{https://surfer.nmr.mgh.harvard.edu/fswiki/CoordinateSystems}
\code{\link{MNI152_to_MNI305}}, \code{\link{transform_coords}}
}
\keyword{datasets}
