% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olsen_mtl.R
\docType{data}
\name{olsen_mtl}
\alias{olsen_mtl}
\title{Medial Temporal Lobe atlas from Rosanna Olsen}
\format{
A list of the 'atlas' class with 7 elements
\describe{
  \item{name}{the name of the atlas}
  \item{atlas}{the atlas in MNI space with 1X1X1 resolution }
  \item{labels}{the atlas labels}
  \item{orig_labels}{the full labels (including hemisphere)}
  \item{ids}{the integer ids from 1 to 16}
  \item{hemi}{the hemisphere of each region}
}
}
\usage{
olsen_mtl
}
\description{
Medial Temporal Lobe atlas from Rosanna Olsen
}
\examples{
data(olsen_mtl)
}
\keyword{datasets}
