% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_data.R
\name{parcel_values}
\alias{parcel_values}
\title{Extract Parcel Values Aligned to an Atlas}
\usage{
parcel_values(x, atlas, column = "value")
}
\arguments{
\item{x}{A `parcel_data` object.}

\item{atlas}{An atlas object.}

\item{column}{Value column in `x$parcels` to extract.}
}
\value{
A vector with `length(atlas$ids)` elements ordered to `atlas$ids`.
}
\description{
Returns a vector aligned to `atlas$ids`, suitable for `map_atlas()` or
`plot_brain()`.
}
