% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_methods.R, R/glasser.R
\name{plot-methods}
\alias{plot-methods}
\alias{plot.atlas}
\alias{plot.glasser}
\title{Plot methods for atlas objects}
\usage{
\method{plot}{atlas}(x, y, ...)
\method{plot}{glasser}(x, y, vals = NULL, thresh = c(0, 0), pos = FALSE,
                       position = "stacked", colour = "gray", guide = TRUE,
                       palette = "cork", lim = NULL, ...)
}
\arguments{
\item{x}{An atlas object}
\item{y}{Unused}
\item{...}{Additional arguments}
}
\description{
S3 \code{plot} methods for visualising atlas data.
}
