% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R, R/glasser.R
\name{print-methods}
\alias{print-methods}
\alias{print.atlas}
\alias{print.glasser}
\title{Print methods for atlas objects}
\usage{
\method{print}{atlas}(x, ...)
\method{print}{glasser}(x, ...)
}
\arguments{
\item{x}{An atlas object}
\item{...}{Additional arguments passed to print methods}
}
\description{
S3 \code{print} methods for objects inheriting from class \code{atlas}.
}
