% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{print.templateflow}
\alias{print.templateflow}
\title{Print a TemplateFlow Object}
\usage{
\method{print}{templateflow}(x, ...)
}
\arguments{
\item{x}{An object of class \code{templateflow}.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Provides a brief summary of the TemplateFlow interface object.
}
\examples{
\donttest{
if (reticulate::py_available(initialize = TRUE) &&
    reticulate::py_module_available("templateflow")) {
  tryCatch({tf <- create_templateflow(); print(tf)},
    error = function(e) NULL)
}
}
}
