% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/atlas.R
\name{reduce_atlas}
\alias{reduce_atlas}
\alias{reduce_atlas.atlas}
\title{Reduce a NeuroVol or NeuroVec by an Atlas}
\usage{
reduce_atlas(atlas, data_vol, stat_func, ...)

\method{reduce_atlas}{atlas}(atlas, data_vol, stat_func, ...)
}
\arguments{
\item{atlas}{An atlas object or another object for which a method is defined.}

\item{data_vol}{A \code{NeuroVol} (3D) or \code{NeuroVec} (4D) object containing the data
to be summarized.}

\item{stat_func}{The function to apply to the data within each ROI (e.g., \code{mean},
\code{sd}, \code{sum}).}

\item{...}{Additional arguments to be passed to \code{stat_func} or other methods.}
}
\value{
A \code{tibble} summarizing the data_vol by atlas regions.
}
\description{
Applies a summary function to data within each ROI defined by an atlas.
This is an S3 generic function.
}
