% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{show_templateflow_cache_path}
\alias{show_templateflow_cache_path}
\title{Show neuroatlas TemplateFlow Cache Path}
\usage{
show_templateflow_cache_path()
}
\value{
A character string representing the path to the TemplateFlow cache directory.
}
\description{
Returns the path to the `neuroatlas` package's cache directory used for 
TemplateFlow downloads. This is typically located within the path returned by
`tools::R_user_dir("neuroatlas", "cache")`, in a subdirectory named "templateflow".
}
\examples{
cat("TemplateFlow cache is at:", show_templateflow_cache_path(), "\n")
}
