% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_flow.R
\name{templates}
\alias{templates}
\title{List Available Templates}
\usage{
templates()
}
\value{
A character vector of available template names
}
\description{
Returns a list of all available templates in the Templateflow repository.
}
\keyword{internal}
